cleaned up testting fixtures
